+++ packages/aws_common/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Bump minimum Dart SDK to 2.17
 - Consolidate `AWSHttpMethodHelper` into `AWSHttpMethod`
 - Add `AWSDebuggable` mixin for debug-friendly `toString`
+- Add `AWSLogger` class and utilities
 
 ## 0.1.0 (2022-05-17)